FloraWatch/backend/requirements.txt
# Dependencias principales
flask==2.3.3
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
folium==0.14.0
scikit-learn==1.3.0
plotly==5.15.0

# Procesamiento geoespacial
rasterio==1.3.8
geopandas==0.13.2
shapely==2.0.1
pyproj==3.6.0
fiona==1.9.4

# APIs y web scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilidades
python-dotenv==1.0.0
pillow==10.0.0
tqdm==4.66.1
joblib==1.3.2

# Base de datos
sqlite3

# Servidor web
gunicorn==21.2.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Desarrollo
black==23.7.0
flake8==6.0.0